ref(kafka): Create statsd module
